#17 Extract SocialNav into component
